requirements.txt
pytest
requests
allure-pytest==2.15.0
allure-python-commons==2.15.0
pydantic